+++ example/fft_demo/main.c
@@ -1,10 +1,14 @@
-// Copyright 2020-2021 XMOS LIMITED.
+// Copyright 2020-2022 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+
+#ifdef __XS3A__
+# include <xscope.h>
+#endif
 
 #include "bfp_math.h"
 
@@ -16,6 +20,9 @@ void fft_complex_example();
 
 int main(int argc, char** argv)
 {
+#ifdef __XS3A__
+  xscope_config_io(XSCOPE_IO_BASIC);
+#endif
 
   // Seed the random number generator, using a constant for reproducibility
   srand(RAND_SEED);