+++ lib_xs3_math/api/xs3_vpu_scalar_ops.h
@@ -296,6 +296,15 @@ int32_t vcmci32(
 
 
 
+/**
+ * Implements the logic of the FMACC instruction for IEEE754 single-precision
+ * floats.
+ */
+float fmacc(
+    const float acc,
+    const float b,
+    const float c);
+
 
 
 #endif //XS3_VPU_SCALAR_OPS_H_