+++ test/fft_tests/src/main.c
@@ -1,8 +1,12 @@
-// Copyright 2020-2021 XMOS LIMITED.
+// Copyright 2020-2022 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 
 #include <stdio.h>
+
+#ifdef __XS3A__
+# include <xscope.h>
+#endif
 
 #include "unity_fixture.h"
 #include "tst_common.h"
@@ -11,26 +15,32 @@ FILE* perf_file = NULL;
 
 int main(int argc, const char* argv[])
 {
-    UnityGetCommandLineOptions(argc, argv);
-    UnityBegin(argv[0]);
+#ifdef __XS3A__
+  xscope_config_io(XSCOPE_IO_BASIC);
+#endif
+    
+  UnityGetCommandLineOptions(argc, argv);
+  UnityBegin(argv[0]);
 
 #if WRITE_PERFORMANCE_INFO
-    printf("Writing performance info to: %s\n", PERFORMANCE_INFO_FILENAME);
-    perf_file = fopen(PERFORMANCE_INFO_FILENAME, "w");
-    fprintf(perf_file, "Function, Input Size, Worst Observed Error, Worst Observed Timing, Misc\n");
+  printf("Writing performance info to: %s\n", PERFORMANCE_INFO_FILENAME);
+  perf_file = fopen(PERFORMANCE_INFO_FILENAME, "w");
+  fprintf(perf_file, "Function, Input Size, Worst Observed Error, Worst Observed Timing, Misc\n");
 #endif
 
-    RUN_TEST_GROUP(xs3_fft_helpers);
-    RUN_TEST_GROUP(xs3_fft_mono_adjust);
-    RUN_TEST_GROUP(xs3_fft_dit);
-    RUN_TEST_GROUP(xs3_fft_dif);
+  RUN_TEST_GROUP(xs3_fft_helpers);
+  RUN_TEST_GROUP(xs3_fft_mono_adjust);
+  RUN_TEST_GROUP(xs3_fft_dit);
+  RUN_TEST_GROUP(xs3_fft_dif);
 
-    RUN_TEST_GROUP(bfp_fft);
-    RUN_TEST_GROUP(bfp_fft_packing);
+  RUN_TEST_GROUP(bfp_fft);
+  RUN_TEST_GROUP(bfp_fft_packing);
+  
+  RUN_TEST_GROUP(xs3_vect_f32_fft);
 
 #if WRITE_PERFORMANCE_INFO
-    fclose(perf_file);
+  fclose(perf_file);
 #endif
 
-    return UNITY_END();
+  return UNITY_END();
 }