+++ test/scalar_tests/src/main.c
@@ -1,13 +1,21 @@
-// Copyright 2020-2021 XMOS LIMITED.
+// Copyright 2020-2022 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 
 #include <stdio.h>
 
+#ifdef __XS3A__
+# include <xscope.h>
+#endif
+
 #include "unity_fixture.h"
 
 int main(int argc, const char* argv[])
 {
+#ifdef __XS3A__
+  xscope_config_io(XSCOPE_IO_BASIC);
+#endif
+    
     UnityGetCommandLineOptions(argc, argv);
     UnityBegin(argv[0]);